+++ dsfhub/resource_secret_manager.go
@@ -552,7 +552,7 @@ func resourceSecretManagerRead(d *schema.ResourceData, m interface{}) error {
 	if secretManagerReadResponse.Data.AssetData.ServerPort != nil {
 		var serverPort string
 		if serverPortNum, ok := secretManagerReadResponse.Data.AssetData.ServerPort.(float64); ok {
-			serverPort = fmt.Sprintf("%d", float64(serverPortNum))
+			serverPort = fmt.Sprintf("%d", int(serverPortNum))
 		} else {
 			serverPort = secretManagerReadResponse.Data.AssetData.ServerPort.(string)
 		}
